Create Phys 172 Space Shuttle on Mars Simulation
P.S this is Vpython code and not Python code. I would submit an old python code from my high school classes, but those got erased when I had to turn in my school provided laptop at the end of 12th grade back in 2020, so run this in Glowscript to get it to work.